+++ python/bartons/typing.py
@@ -1,7 +1,7 @@
 import sys
 import polars as pl
 
-from typing import Union
+from typing import Literal, Union
 
 if sys.version_info >= (3, 10):
     from typing import TypeAlias
@@ -10,6 +10,7 @@ else:
 from polars.datatypes import DataType, DataTypeClass
 
 IntoExprColumn: TypeAlias = Union[pl.Expr, str, pl.Series]
+MAType: TypeAlias = Literal["sma", "ema", "wma", "dema", "tema", "kama"]
 PolarsDataType: TypeAlias = Union[DataType, DataTypeClass]